# api/openapi.yaml
openapi: 3.1.0
info:
  title: ConsorcioAbierto API
  version: 0.1.0
  description: |
    API del SaaS multi-tenant de administración de consorcios.
    Contrato de Fase 0. Convenciones: JSON snake_case, fechas ISO 8601,
    importes como enteros `*_cents` + `currency`, IDs UUID.
    Errores RFC 9457 (application/problem+json).
servers:
  - url: /api/v1
security:
  - cookieAuth: []
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: __Host-ca_session
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: false
      schema:
        type: string
      description: Obligatorio en confirmar, publicar, cobranzas, asignaciones y webhooks.
...
    ConsorcioId:
      name: id
      in: path
      required: true
      schema: { type: string, format: uuid }
    UnidadId:
      name: id
      in: path
      required: true
      schema: { type: string, format: uuid }
...
    ComunicadoId:
      name: id
      in: path
      required: true
      schema: { type: string, format: uuid }
    ReclamoId:
      name: id
      in: path
      required: true
      schema: { type: string, format: uuid }
...
  headers:
    RequestIdOut:
      schema:
        type: string
        format: uuid
  responses:
    Problem:
      description: Error RFC 9457.
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
  schemas:
...
    ErrorField:
      type: object
      required: [field, code]
      properties:
        field:
          type: string
        code:
          type: string
    Problem:
      type: object
      required: [type, title, status, code, detail, instance, request_id]
      properties:
        type:
          type: string
          format: uri
        title:
          type: string
        status:
          type: integer
        code:
          type: string
        detail:
          type: string
        instance:
          type: string
        request_id:
          type: string
          format: uuid
        errors:
          type: array
          items:
            $ref: "#/components/schemas/ErrorField"
    Page:
      type: object
      required: [data, meta]
      properties:
        data:
          type: array
          items: {}
        meta:
          type: object
          required: [request_id]
          properties:
            request_id:
              type: string
              format: uuid
            page:
              type: integer
            per_page:
              type: integer
            total:
              type: integer
    Cursor:
      type: object
      required: [data, meta]
      properties:
        data:
          type: array
          items: {}
        meta:
          type: object
          required: [request_id, next_cursor]
          properties:
            request_id:
              type: string
              format: uuid
            next_cursor:
              type: [string, "null"]
...
    PortalConsorcioSummary:
      type: object
      required: [id, nombre, estado, saldo_vencido_cents, cargos_vencidos, vencido_desde]
      properties:
        id: { type: string, format: uuid }
        nombre: { type: string }
        estado: { type: string }
        saldo_vencido_cents:
          type: integer
          format: int64
        cargos_vencidos:
          type: integer
          format: int64
        vencido_desde:
          type: string
    PortalReceiptSummary:
      type: object
      required: [consorcio_id, consorcio_nombre, cobranza_id, fecha, importe_cents, estado, saldo_a_favor_cents, created_at]
      properties:
        consorcio_id: { type: string, format: uuid }
        consorcio_nombre: { type: string }
        cobranza_id: { type: string, format: uuid }
        fecha: { type: string }
        importe_cents:
          type: integer
          format: int64
        estado: { type: string }
        saldo_a_favor_cents:
          type: integer
          format: int64
        created_at:
          type: string
          description: RFC3339 timestamp.
    PortalComunicadoSummary:
      type: object
      required: [id, consorcio_id, consorcio_nombre, titulo, estado, publicado_at]
      properties:
        id: { type: string, format: uuid }
        consorcio_id: { type: string, format: uuid }
        consorcio_nombre: { type: string }
        titulo: { type: string }
        estado: { type: string }
        publicado_at:
          type: string
          description: RFC3339 timestamp.
    PortalReclamoSummary:
      type: object
      required: [id, consorcio_id, unidad_id, categoria, estado, texto, created_at]
      properties:
        id: { type: string, format: uuid }
        consorcio_id: { type: string, format: uuid }
        unidad_id: { type: string, format: uuid }
        categoria: { type: string }
        estado: { type: string }
        texto: { type: string }
        created_at:
          type: string
          description: RFC3339 timestamp.
    PortalHome:
      type: object
      required: [consorcios, recibos_recientes, comunicados_recientes, reclamos_recientes, total_saldo_vencido_cents]
      properties:
        consorcios:
          type: array
          items:
            $ref: "#/components/schemas/PortalConsorcioSummary"
        recibos_recientes:
          type: array
          items:
            $ref: "#/components/schemas/PortalReceiptSummary"
        comunicados_recientes:
          type: array
          items:
            $ref: "#/components/schemas/PortalComunicadoSummary"
        reclamos_recientes:
          type: array
          items:
            $ref: "#/components/schemas/PortalReclamoSummary"
        total_saldo_vencido_cents:
          type: integer
          format: int64
...
    Reclamo:
      type: object
      required: [id, unidad_id, categoria, estado]
      properties:
        id: { type: string, format: uuid }
        unidad_id: { type: string, format: uuid }
        categoria: { type: string }
        estado:
          type: string
          enum: [abierto, en_progreso, resuelto, cerrado]
        responsable_id: { type: [string, "null"], format: uuid }
        created_at: { type: string, format: date-time }
    ReclamoMensaje:
      type: object
      required: [id, reclamo_id, texto, created_at]
      properties:
        id: { type: string, format: uuid }
        reclamo_id: { type: string, format: uuid }
        texto: { type: string }
        adjunto_id: { type: [string, "null"], format: uuid }
        created_at: { type: string, format: date-time }
    ReclamoTransicion:
      type: object
      required: [id, reclamo_id, accion, created_at]
      properties:
        id: { type: string, format: uuid }
        reclamo_id: { type: string, format: uuid }
        accion: { type: string }
        motivo: { type: [string, "null"] }
        responsable_id: { type: [string, "null"], format: uuid }
        created_at: { type: string, format: date-time }
    ReclamoDetalle:
      type: object
      required: [reclamo, mensajes, transiciones]
      properties:
        reclamo:
          $ref: "#/components/schemas/Reclamo"
        mensajes:
          type: array
          items:
            $ref: "#/components/schemas/ReclamoMensaje"
        transiciones:
          type: array
          items:
            $ref: "#/components/schemas/ReclamoTransicion"
    Comunicado:
      type: object
      required: [id, titulo, estado]
      properties:
        id: { type: string, format: uuid }
        titulo: { type: string }
        estado:
          type: string
          enum: [borrador, publicado]
        destinatarios: { type: integer }
...
paths:
...
  /portal:
    get:
      operationId: getPortalHome
      tags: [auth]
      responses:
        "200":
          description: Resumen del portal del tenant activo.
          headers:
            X-Request-ID:
              $ref: "#/components/headers/RequestIdOut"
          content:
            application/json:
              schema:
                type: object
                required: [data, meta]
                properties:
                  data:
                    $ref: "#/components/schemas/PortalHome"
                  meta:
                    type: object
                    required: [request_id]
                    properties:
                      request_id:
                        type: string
                        format: uuid
        "401": { $ref: "#/components/responses/Problem" }
...
  /unidades/{id}/cuenta-corriente:
    parameters:
      - { $ref: "#/components/parameters/UnidadId" }
    get:
      operationId: getCuentaCorriente
      tags: [cobranzas]
      responses:
        "200":
          description: Movimientos y saldo reconstruido de la UF.
          content:
            application/json:
              schema:
                type: object
                properties:
                  saldo_cents: { type: integer, format: int64 }
                  saldo_a_favor_cents: { type: integer, format: int64 }
                  movimientos:
                    $ref: "#/components/schemas/Cursor"
...
  /consorcios/{id}/morosidad:
    parameters:
      - { $ref: "#/components/parameters/ConsorcioId" }
    get:
      operationId: getMorosidad
      tags: [cobranzas]
      responses:
        "200":
          description: Resumen de deuda vencida por UF.
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      type: object
                      properties:
                        unidad_id: { type: string, format: uuid }
                        unidad_codigo: { type: string }
                        saldo_vencido_cents: { type: integer, format: int64 }
                        cantidad_cargos: { type: integer, format: int64 }
                        vencido_desde: { type: string, format: date }
                  meta:
                    type: object
                    properties:
                      request_id: { type: string, format: uuid }
                      total_saldo_cents: { type: integer, format: int64 }
                      total_cargos_vencidos: { type: integer, format: int64 }
        "400": { $ref: "#/components/responses/Problem" }
...
  /consorcios/{id}/unidades/{unidadId}/cuenta-corriente:
    parameters:
      - { $ref: "#/components/parameters/ConsorcioId" }
      - { $ref: "#/components/parameters/UnidadId" }
    get:
      operationId: getCuentaCorriente
      tags: [cobranzas]
      responses:
        "200":
          description: Libro de movimientos de una UF.
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      type: object
                      properties:
                        id: { type: string, format: uuid }
                        unidad_id: { type: string, format: uuid }
                        tipo: { type: string }
                        fecha_efectiva: { type: string, format: date }
                        debit_cents: { type: integer, format: int64 }
                        credit_cents: { type: integer, format: int64 }
                        currency: { type: string }
                        referencia: { type: [string, "null"] }
                  meta:
                    type: object
                    properties:
                      request_id: { type: string, format: uuid }
                      saldo_cents: { type: integer, format: int64 }
        "404": { $ref: "#/components/responses/Problem" }
        "400": { $ref: "#/components/responses/Problem" }
...
  /consorcios/{id}/comunicados:
    parameters:
      - { $ref: "#/components/parameters/ConsorcioId" }
    get:
      operationId: listComunicados
      tags: [comunicaciones]
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Page"
    post:
      operationId: createComunicado
      tags: [comunicaciones]
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [titulo, cuerpo]
              properties:
                titulo: { type: string }
                cuerpo: { type: string }
                destinatarios:
                  type: string
                  enum: [todos, unidades]
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Comunicado"
  /comunicados/{id}/publicar:
    parameters:
      - { $ref: "#/components/parameters/ComunicadoId" }
    post:
      operationId: publicarComunicado
      tags: [comunicaciones]
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      responses:
        "202":
          description: Envíos encolados en el worker.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Comunicado"
  /consorcios/{id}/reclamos:
    parameters:
      - { $ref: "#/components/parameters/ConsorcioId" }
    get:
      operationId: listReclamos
      tags: [reclamos]
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Page"
    post:
      operationId: createReclamo
      tags: [reclamos]
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [unidad_id, categoria, texto]
              properties:
                unidad_id: { type: string, format: uuid }
                categoria: { type: string }
                texto: { type: string, minLength: 3 }
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Reclamo"
        "403": { $ref: "#/components/responses/Problem" }
  /reclamos/{id}:
    parameters:
      - { $ref: "#/components/parameters/ReclamoId" }
    get:
      operationId: getReclamo
      tags: [reclamos]
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReclamoDetalle"
  /reclamos/{id}/mensajes:
    parameters:
      - { $ref: "#/components/parameters/ReclamoId" }
    post:
      operationId: addReclamoMensaje
      tags: [reclamos]
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [texto]
              properties:
                texto: { type: string, minLength: 1 }
                adjunto_id: { type: [string, "null"], format: uuid }
      responses:
        "201": { description: Mensaje agregado. }
...